Added new questions

+++ src/main/java/org/java_tutorials/util/Hello.java
@@ -7,6 +7,6 @@ import java.util.*;
 public class Hello {
     public static void main(String[] args) {
         System.out.println(Util.findMax(
-                new Integer[]{5,9,4,3}));
+                new Double[]{5.0,9.0,4.0,3.0}));
     }
 }
